Add checksum to the LZ4 compressed buffer format.
Simply reuses the XXHASH implementation from the LZ4 library (also
used in the LZ4 frame format).
